kitchenassistant-application/src/main/kotlin/pl/szczeliniak/kitchenassistant/shoppinglist/commands/DeleteShoppingListItemCommand.kt
package pl.szczeliniak.kitchenassistant.shoppinglist.commands

import pl.szczeliniak.kitchenassistant.shared.ErrorCode
import pl.szczeliniak.kitchenassistant.shared.KitchenAssistantException
import pl.szczeliniak.kitchenassistant.shared.dtos.SuccessResponse
import pl.szczeliniak.kitchenassistant.shoppinglist.ShoppingListDao
import pl.szczeliniak.kitchenassistant.shoppinglist.ShoppingListItemDao

class DeleteShoppingListItemCommand(
    private val shoppingListDao: ShoppingListDao,
    private val shoppingListItemDao: ShoppingListItemDao
) {

    fun execute(shoppingListId: Int, shoppingListItemId: Int): SuccessResponse {
        val shoppingList =
            shoppingListDao.findById(shoppingListId)
                ?: throw KitchenAssistantException(ErrorCode.SHOPPING_LIST_NOT_FOUND)

        val item = shoppingList.items.firstOrNull { it.id == shoppingListItemId }
            ?: throw KitchenAssistantException(ErrorCode.SHOPPING_LIST_ITEM_NOT_FOUND)

        if (item.deleted) {
            throw KitchenAssistantException(ErrorCode.SHOPPING_LIST_ITEM_ALREADY_REMOVED)
        }
        item.deleted = true

        shoppingListItemDao.save(item)
        return SuccessResponse(shoppingList.id)
    }

}